Move callable factories to new JobQueueConfig
The JobQueue\Config class has gotten large and I
want to split its functionality into smaller pieces
