+++ tests/src/Hodor/JobQueue/ConfigTest.php
@@ -114,6 +114,44 @@ class ConfigTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::__construct
+     * @covers ::getJobQueueConfig
+     * @dataProvider configProvider
+     */
+    public function testJobQueueConfigIsReused($options)
+    {
+        $config = new Config(__FILE__, $options);
+
+        $this->assertSame($config->getJobQueueConfig(), $config->getJobQueueConfig());
+    }
+
+    /**
+     * @covers ::__construct
+     * @covers ::getJobQueueConfig
+     * @dataProvider configProvider
+     */
+    public function testJobQueueConfigOptionsArePassedIn($options)
+    {
+        $config = new Config(__FILE__, $options);
+
+        $job_queue_config = $config->getJobQueueConfig();
+
+        $uniqid = uniqid();
+        $this->assertSame(
+            $uniqid,
+            call_user_func($job_queue_config->getBufferQueueNameFactory(), $uniqid, [], [])
+        );
+        $this->assertSame(
+            $uniqid,
+            call_user_func($job_queue_config->getWorkerQueueNameFactory(), $uniqid, [], [])
+        );
+        $this->assertSame(
+            [$uniqid, ['value' => $uniqid]],
+            call_user_func($job_queue_config->getJobRunnerFactory(), $uniqid, ['value' => $uniqid], [])
+        );
+    }
+
+    /**
+     * @covers ::__construct
      * @covers ::getBufferQueueConfig
      * @dataProvider configProvider
      * @expectedException Exception
@@ -287,188 +325,6 @@ class ConfigTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(
             1,
             $queue_config['fetch_count']
-        );
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getJobRunnerFactory
-     * @expectedException Exception
-     */
-    public function testAJobRunnerFactoryMustBeConfigured()
-    {
-        $config = new Config(__FILE__, []);
-
-        $config->getJobRunnerFactory();
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getJobRunnerFactory
-     * @expectedException Exception
-     */
-    public function testTheJobRunnerFactoryMustBeACallable()
-    {
-        $config = new Config(__FILE__, [
-            'job_runner' => 'blah',
-        ]);
-
-        $config->getJobRunnerFactory();
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getJobRunnerFactory
-     * @dataProvider configProvider
-     */
-    public function testTheJobRunnerFactoryIsReturnedIfProperlyConfigured($options)
-    {
-        $config = new Config(__FILE__, $options);
-
-        $callback = $config->getJobRunnerFactory();
-
-        $this->assertTrue(is_callable($callback));
-        $this->assertSame(
-            $options['job_runner'],
-            $callback
-        );
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getWorkerQueueNameFactory
-     * @expectedException \Exception
-     */
-    public function testWorkerQueueNameFactoryThrowsExceptionIfItIsNotCallable()
-    {
-        $config = new Config(__FILE__, [
-            'worker_queue_name_factory' => 'blah',
-        ]);
-
-        $config->getWorkerQueueNameFactory();
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getWorkerQueueNameFactory
-     * @dataProvider configProvider
-     */
-    public function testWorkerQueueNameFactoryIsDefaultedToQueueNameOptionsCallback($options)
-    {
-        unset($options['worker_queue_name_factory']);
-        $config = new Config(__FILE__, $options);
-
-        $callback = $config->getWorkerQueueNameFactory();
-
-        $this->assertTrue(is_callable($callback));
-        $this->assertEquals(
-            'default',
-            call_user_func($callback, 'n/a', [], ['queue_name' => 'default'])
-        );
-        $this->assertEquals(
-            'other',
-            call_user_func($callback, 'n/a', [], ['queue_name' => 'other'])
-        );
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getWorkerQueueNameFactory
-     * @dataProvider configProvider
-     * @expectedException \Exception
-     */
-    public function testDefaultWorkerQueueNameThrowsAnExceptionIfQueueNameIsNotProvided($options)
-    {
-        unset($options['worker_queue_name_factory']);
-        $config = new Config(__FILE__, $options);
-
-        $callback = $config->getWorkerQueueNameFactory();
-
-        $this->assertTrue(is_callable($callback));
-        $this->assertEquals(
-            'other',
-            call_user_func($callback, 'n/a', [], ['not_queue_name' => 'other'])
-        );
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getWorkerQueueNameFactory
-     * @dataProvider configProvider
-     */
-    public function testWorkerQueueNameFactoryCanBeProvided($options)
-    {
-        $config = new Config(__FILE__, $options);
-
-        $callback = $config->getWorkerQueueNameFactory();
-
-        $this->assertTrue(is_callable($callback));
-        $this->assertEquals(
-            'non-default',
-            call_user_func($callback, 'non-default', [], ['queue_name' => 'default'])
-        );
-        $this->assertEquals(
-            'another',
-            call_user_func($callback, 'another', [], ['queue_name' => 'other'])
-        );
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getBufferQueueNameFactory
-     * @expectedException \Exception
-     */
-    public function testBufferQueueNameFactoryThrowsExceptionIfItIsNotCallable()
-    {
-        $config = new Config(__FILE__, [
-            'buffer_queue_name_factory' => 'blah',
-        ]);
-
-        $config->getBufferQueueNameFactory();
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getBufferQueueNameFactory
-     * @dataProvider configProvider
-     */
-    public function testBufferQueueNameFactoryIsDefaultedToDefaultQueue($options)
-    {
-        unset($options['buffer_queue_name_factory']);
-        $config = new Config(__FILE__, $options);
-
-        $callback = $config->getBufferQueueNameFactory();
-
-        $this->assertTrue(is_callable($callback));
-        $this->assertEquals(
-            'default',
-            call_user_func($callback, 'n/a', [], ['queue_name' => 'default'])
-        );
-        $this->assertEquals(
-            'default',
-            call_user_func($callback, 'n/a', [], ['queue_name' => 'other'])
-        );
-    }
-
-    /**
-     * @covers ::__construct
-     * @covers ::getBufferQueueNameFactory
-     * @dataProvider configProvider
-     */
-    public function testBufferQueueNameFactoryCanBeProvided($options)
-    {
-        $config = new Config(__FILE__, $options);
-
-        $callback = $config->getBufferQueueNameFactory();
-
-        $this->assertTrue(is_callable($callback));
-        $this->assertEquals(
-            'non-default',
-            call_user_func($callback, 'non-default', [], ['queue_name' => 'default'])
-        );
-        $this->assertEquals(
-            'another',
-            call_user_func($callback, 'another', [], ['queue_name' => 'other'])
         );
     }
